Add initial refactor of logic classes into usecases
