+++ src/components/mapping-table/MappingTable.tsx
@@ -17,13 +17,15 @@ import MappingWizard, {
     prepareSteps,
 } from "../../components/mapping-wizard/MappingWizard";
 import MetadataTable from "../../components/metadata-table/MetadataTable";
+import { useAppContext } from "../../contexts/ApiContext";
+import { cleanOrgUnitPath } from "../../domain/synchronization/utils";
 import { D2Model } from "../../models/dhis/default";
 import { d2ModelFactory } from "../../models/dhis/factory";
 import { ProgramDataElementModel } from "../../models/dhis/mapping";
 import { DataElementModel, OrganisationUnitModel } from "../../models/dhis/metadata";
 import Instance, { MetadataMapping, MetadataMappingDictionary } from "../../models/instance";
+import { D2 } from "../../types/d2";
 import { MetadataType } from "../../utils/d2";
-import { cleanOrgUnitPath } from "../../utils/synchronization";
 import {
     autoMap,
     buildDataElementFilterForProgram,
@@ -32,8 +34,6 @@ import {
     EXCLUDED_KEY,
     getChildrenRows,
 } from "./utils";
-import { useAppContext } from "../../contexts/ApiContext";
-import { D2 } from "../../types/d2";
 
 const useStyles = makeStyles({
     iconButton: {
