+++ src/domain/aggregated/entities/Aggregated.ts
@@ -1,0 +1,17 @@
+export interface DataValue {
+    dataElement: string;
+    period: string;
+    orgUnit: string;
+    categoryOptionCombo: string;
+    attributeOptionCombo?: string;
+    value: string;
+    storedBy: string;
+    created: string;
+    lastUpdated: string;
+    followUp: boolean;
+    comment?: string;
+}
+
+export interface AggregatedPackage {
+    dataValues: DataValue[];
+}
