+++ src/domain/metadata/usecases/MetadataSyncUseCase.ts
@@ -1,30 +1,29 @@
 import _ from "lodash";
 import memoize from "nano-memoize";
-
-import InstanceEntity from "../../instance/Instance";
-import { GenericSync, SyncronizationPayload } from "../../../logic/sync/generic";
 import MetadataD2ApiRepository from "../../../data/metadata/repositories/MetadataD2ApiRepository";
-import { MetadataRepository } from "../MetadataRepositoriy";
-import { MetadataPackage, MetadataEntities } from "../entities/MetadataEntities";
-import { Ref } from "../../common/entities/Schemas";
-import {
-    buildNestedRules,
-    cleanObject,
-    getAllReferences,
-    cleanReferences,
-    cleanMetadataImportResponse,
-} from "../utils";
-
-//TODO: Uncouple this dependencies. This class should be moved to domain
-// and It should have not any dependency outside from the domain
-import { D2Api } from "../../../types/d2-api";
-import { D2 } from "../../../types/d2";
-import Instance from "../../../models/instance";
 import { d2ModelFactory } from "../../../models/dhis/factory";
+import Instance from "../../../models/instance";
+import { D2 } from "../../../types/d2";
+import { D2Api } from "../../../types/d2-api";
 import { ExportBuilder, NestedRules, SynchronizationBuilder } from "../../../types/synchronization";
 import { promiseMap } from "../../../utils/common";
+import { Ref } from "../../common/entities/Schemas";
+import InstanceEntity from "../../instance/Instance";
+import {
+    GenericSyncUseCase,
+    SyncronizationPayload,
+} from "../../synchronization/usecases/GenericSyncUseCase";
+import { MetadataEntities, MetadataPackage } from "../entities/MetadataEntities";
+import { MetadataRepository } from "../MetadataRepositoriy";
+import {
+    buildNestedRules,
+    cleanMetadataImportResponse,
+    cleanObject,
+    cleanReferences,
+    getAllReferences,
+} from "../utils";
 
-export class MetadataSyncUseCase extends GenericSync {
+export class MetadataSyncUseCase extends GenericSyncUseCase {
     public readonly type = "metadata";
     private metadataRepository: MetadataRepository;
