+++ src/utils/synchronization.ts
@@ -1,24 +1,15 @@
 import i18n from "@dhis2/d2-i18n";
-import { AxiosError } from "axios";
-import { D2Api, D2CategoryOptionCombo } from "../types/d2-api";
 import FileSaver from "file-saver";
 import _ from "lodash";
-import moment, { Moment } from "moment";
+import moment from "moment";
 import memoize from "nano-memoize";
-import { SyncronizationClass } from "../logic/sync/generic";
+import { SyncronizationClass } from "../domain/synchronization/usecases/GenericSyncUseCase";
 import Instance, { MetadataMapping, MetadataMappingDictionary } from "../models/instance";
 import SyncRule from "../models/syncRule";
-import { D2, DataImportParams, DataImportResponse, MetadataImportResponse } from "../types/d2";
-import {
-    CategoryOptionAggregationBuilder,
-    DataSyncAggregation,
-    DataSynchronizationParams,
-    SynchronizationResult,
-    SyncRuleType,
-} from "../types/synchronization";
+import { D2, DataImportResponse } from "../types/d2";
+import { D2Api, D2CategoryOptionCombo } from "../types/d2-api";
+import { SynchronizationResult, SyncRuleType } from "../types/synchronization";
 import "../utils/lodash-mixins";
-import { promiseMap } from "./common";
-import { AggregatedPackage, ProgramEvent, DataValue } from "../domain/synchronization/DataEntities";
 
 //TODO: when all request to metadata using metadataRepository.getMetadataByIds
 // this function should be removed
@@ -44,27 +35,6 @@ export async function getMetadata(
     const results = _.deepMerge({}, ...response);
     if (results.system) delete results.system;
     return results;
-}
-
-function buildResponseError(error: AxiosError): MetadataImportResponse {
-    if (error.response && error.response.data) {
-        const {
-            httpStatus = "Unknown",
-            httpStatusCode = 400,
-            message = "Request failed unexpectedly",
-        } = error.response.data;
-        return {
-            ...error.response.data,
-            message: `Error ${httpStatusCode} (${httpStatus}): ${message}`,
-        };
-    } else if (error.response) {
-        const { status, statusText } = error.response;
-        console.error(status, statusText, error);
-        return { status: "ERROR", message: `Unknown error: ${status} ${statusText}` };
-    } else {
-        console.error(error);
-        return { status: "NETWORK ERROR" };
-    }
 }
 
 export function cleanDataImportResponse(
@@ -126,175 +96,6 @@ export const availablePeriods: {
     LAST_FIVE_YEARS: { name: i18n.t("Last 5 years"), start: [5, "year"], end: [1, "year"] },
 };
 
-function buildPeriodFromParams(params: DataSynchronizationParams): [Moment, Moment] {
-    const {
-        period,
-        startDate = "1970-01-01",
-        endDate = moment()
-            .add(1, "years")
-            .endOf("year")
-            .format("YYYY-MM-DD"),
-    } = params;
-
-    if (!period || period === "ALL" || period === "FIXED") {
-        return [moment(startDate), moment(endDate)];
-    } else {
-        const { start, end = start } = availablePeriods[period];
-        if (start === undefined || end === undefined)
-            throw new Error("Unsupported period provided");
-
-        const [startAmount, startType] = start;
-        const [endAmount, endType] = end;
-
-        return [
-            moment()
-                .subtract(startAmount, startType as moment.unitOfTime.DurationConstructor)
-                .startOf(startType as moment.unitOfTime.DurationConstructor),
-            moment()
-                .subtract(endAmount, endType as moment.unitOfTime.DurationConstructor)
-                .endOf(endType as moment.unitOfTime.DurationConstructor),
-        ];
-    }
-}
-
-const aggregations = {
-    DAILY: { format: "YYYYMMDD", unit: "days" as const, amount: 1 },
-    WEEKLY: { format: "YYYY[W]W", unit: "weeks" as const, amount: 1 },
-    MONTHLY: { format: "YYYYMM", unit: "months" as const, amount: 1 },
-    QUARTERLY: { format: "YYYY[Q]Q", unit: "quarters" as const, amount: 1 },
-    YEARLY: { format: "YYYY", unit: "years" as const, amount: 1 },
-};
-
-const buildPeriodsForAggregation = (
-    aggregationType: DataSyncAggregation | undefined,
-    startDate: Moment,
-    endDate: Moment
-): string[] => {
-    if (!aggregationType) return [];
-    const { format, unit, amount } = aggregations[aggregationType];
-
-    const current = startDate.clone();
-    const periods = [];
-
-    while (current.isSameOrBefore(endDate)) {
-        periods.push(current.format(format));
-        current.add(amount, unit);
-    }
-
-    return periods;
-};
-
-export async function getAggregatedData(
-    api: D2Api,
-    params: DataSynchronizationParams,
-    dataSet: string[],
-    dataElementGroup: string[]
-): Promise<AggregatedPackage> {
-    const { orgUnitPaths = [], allAttributeCategoryOptions, attributeCategoryOptions } = params;
-    const [startDate, endDate] = buildPeriodFromParams(params);
-
-    if (dataSet.length === 0 && dataElementGroup.length === 0) return { dataValues: [] };
-
-    const orgUnit = cleanOrgUnitPaths(orgUnitPaths);
-    const attributeOptionCombo = !allAttributeCategoryOptions
-        ? attributeCategoryOptions
-        : undefined;
-
-    return api
-        .get<AggregatedPackage>("/dataValueSets", {
-            dataElementIdScheme: "UID",
-            orgUnitIdScheme: "UID",
-            categoryOptionComboIdScheme: "UID",
-            includeDeleted: false,
-            startDate: startDate.format("YYYY-MM-DD"),
-            endDate: endDate.format("YYYY-MM-DD"),
-            attributeOptionCombo,
-            dataSet,
-            dataElementGroup,
-            orgUnit,
-        })
-        .getData();
-}
-
-export async function getAnalyticsData({
-    api,
-    dataParams,
-    dimensionIds,
-    filter,
-    includeCategories,
-}: {
-    api: D2Api;
-    dataParams: DataSynchronizationParams;
-    dimensionIds: string[];
-    filter?: string[];
-    includeCategories: boolean;
-}): Promise<AggregatedPackage> {
-    const {
-        orgUnitPaths = [],
-        allAttributeCategoryOptions,
-        attributeCategoryOptions,
-        aggregationType,
-    } = dataParams;
-    const [startDate, endDate] = buildPeriodFromParams(dataParams);
-    const periods = buildPeriodsForAggregation(aggregationType, startDate, endDate);
-    const orgUnit = cleanOrgUnitPaths(orgUnitPaths);
-    const attributeOptionCombo = !allAttributeCategoryOptions
-        ? attributeCategoryOptions
-        : undefined;
-
-    if (dimensionIds.length === 0 || orgUnit.length === 0) {
-        return { dataValues: [] };
-    } else if (aggregationType) {
-        const result = await promiseMap(_.chunk(periods, 500), period => {
-            return api
-                .get<AggregatedPackage>("/analytics/dataValueSet.json", {
-                    dimension: _.compact([
-                        `dx:${dimensionIds.join(";")}`,
-                        `pe:${period.join(";")}`,
-                        `ou:${orgUnit.join(";")}`,
-                        includeCategories ? `co` : undefined,
-                        attributeOptionCombo ? `ao:${attributeOptionCombo.join(";")}` : undefined,
-                    ]),
-                    filter,
-                })
-                .getData();
-        });
-
-        const dataValues = _(result)
-            .map(({ dataValues }) => dataValues)
-            .flatten()
-            .compact()
-            .value();
-
-        return { dataValues };
-    } else {
-        throw new Error("Aggregated syncronization requires a valid aggregation type");
-    }
-}
-
-export const getDefaultIds = memoize(
-    async (api: D2Api, filter?: string) => {
-        const response = (await api
-            .get("/metadata", {
-                filter: "code:eq:default",
-                fields: "id",
-            })
-            .getData()) as {
-            [key: string]: { id: string }[];
-        };
-
-        const metadata = _.pickBy(response, (_value, type) => !filter || type === filter);
-
-        return _(metadata)
-            .omit(["system"])
-            .values()
-            .flatten()
-            .map(({ id }) => id)
-            .value();
-    },
-    { serializer: (api: D2Api) => api.baseUrl }
-);
-
 // TODO: when all request to metadata using metadataRepository.getModelByType
 // this function should be removed
 export const getCategoryOptionCombos = memoize(
@@ -315,220 +116,6 @@ export const getCategoryOptionCombos = memoize(
     },
     { serializer: (api: D2Api) => api.baseUrl }
 );
-
-export const getAllDimensions = memoize(
-    async (api: D2Api) => {
-        const { dimensions } = await api
-            .get<{ dimensions: Array<{ id: string }> }>("/dimensions", {
-                paging: false,
-                fields: "id",
-            })
-            .getData();
-
-        return dimensions.map(({ id }) => id);
-    },
-    { serializer: (api: D2Api) => api.baseUrl }
-);
-
-/**
- * Given all the aggregatedDataElements compile a list of dataElements
- * that have aggregation for their category options
- * @param MetadataMappingDictionary
- */
-export const getAggregatedOptions = async (
-    api: D2Api,
-    { aggregatedDataElements }: MetadataMappingDictionary,
-    categoryOptionCombos: Partial<D2CategoryOptionCombo>[]
-): Promise<CategoryOptionAggregationBuilder[]> => {
-    const dimensions = await getAllDimensions(api);
-    const findOptionCombo = (mappedOption: string, mappedCombo?: string) =>
-        categoryOptionCombos.find(
-            ({ categoryCombo, categoryOptions }) =>
-                categoryCombo?.id === mappedCombo &&
-                categoryOptions?.map(({ id }) => id).includes(mappedOption)
-        )?.id ?? mappedOption;
-
-    const validOptions = _.transform(
-        aggregatedDataElements,
-        (result, { mapping = {} }, dataElement) => {
-            const { categoryOptions, categoryCombos } = mapping;
-
-            const builders = _(categoryOptions)
-                .mapValues(({ mappedId = "DISABLED" }, categoryOption) => ({
-                    categoryOption,
-                    mappedId,
-                }))
-                .values()
-                .groupBy(({ mappedId }) => mappedId)
-                .pickBy((values, mappedId) => values.length > 1 && mappedId !== "DISABLED")
-                .mapValues((values = [], mappedCategoryOption) => ({
-                    dataElement,
-                    categoryOptions: values.map(({ categoryOption }) => categoryOption),
-                    mappedOptionCombo: findOptionCombo(
-                        mappedCategoryOption,
-                        _.values(categoryCombos)[0]?.mappedId
-                    ),
-                }))
-                .values()
-                .value();
-            result.push(...builders);
-        },
-        [] as Omit<CategoryOptionAggregationBuilder, "category">[]
-    );
-
-    const result = _.flatten(
-        dimensions.map(category => validOptions.map(item => ({ ...item, category })))
-    );
-
-    return result;
-};
-
-// TODO: when all request to this use metadataRepository.getModelByType
-// this function should be removed
-export const getRootOrgUnit = memoize(
-    (api: D2Api) => {
-        return api.models.organisationUnits.get({
-            filter: { level: { eq: "1" } },
-            fields: { $owner: true },
-        });
-    },
-    { serializer: (api: D2Api) => api.baseUrl }
-);
-
-export function cleanObjectDefault(object: ProgramEvent, defaults: string[]): ProgramEvent;
-export function cleanObjectDefault(object: DataValue, defaults: string[]): DataValue;
-export function cleanObjectDefault(object: ProgramEvent | DataValue, defaults: string[]) {
-    return _.pickBy(object, value => !defaults.includes(String(value)));
-}
-
-export function cleanOrgUnitPath(orgUnitPath?: string): string {
-    return (
-        _(orgUnitPath)
-            .split("/")
-            .last() ?? ""
-    );
-}
-
-export function cleanOrgUnitPaths(orgUnitPaths: string[]): string[] {
-    return orgUnitPaths.map(cleanOrgUnitPath);
-}
-
-export async function getEventsData(
-    api: D2Api,
-    params: DataSynchronizationParams,
-    programs: string[] = []
-) {
-    const { period, orgUnitPaths = [], events = [], allEvents } = params;
-    const [startDate, endDate] = buildPeriodFromParams(params);
-    const defaults = await getDefaultIds(api);
-
-    if (programs.length === 0) return [];
-
-    const orgUnits = cleanOrgUnitPaths(orgUnitPaths);
-
-    const result = [];
-
-    for (const program of programs) {
-        const { events: response } = (await api
-            .get("/events", {
-                paging: false,
-                program,
-                startDate: period !== "ALL" ? startDate.format("YYYY-MM-DD") : undefined,
-                endDate: period !== "ALL" ? endDate.format("YYYY-MM-DD") : undefined,
-            })
-            .getData()) as { events: (ProgramEvent & { event: string })[] };
-
-        result.push(...response);
-    }
-
-    return _(result)
-        .filter(({ orgUnit }) => orgUnits.includes(orgUnit))
-        .filter(({ event }) => (allEvents ? true : events.includes(event)))
-        .map(object => ({ ...object, id: object.event }))
-        .map(object => cleanObjectDefault(object, defaults))
-        .value();
-}
-
-export async function postData(
-    instance: Instance,
-    endpoint: "/events" | "/dataValueSets",
-    data: object,
-    additionalParams: DataImportParams
-): Promise<any> {
-    try {
-        const response = await instance
-            .getApi()
-            .post(
-                endpoint,
-                {
-                    idScheme: "UID",
-                    dataElementIdScheme: "UID",
-                    orgUnitIdScheme: "UID",
-                    eventIdScheme: "UID",
-                    preheatCache: false,
-                    skipExistingCheck: false,
-                    format: "json",
-                    async: false,
-                    dryRun: false,
-                    ...additionalParams,
-                },
-                data
-            )
-            .getData();
-
-        return response;
-    } catch (error) {
-        return buildResponseError(error);
-    }
-}
-
-export async function postAggregatedData(
-    instance: Instance,
-    data: object,
-    additionalParams?: DataImportParams
-): Promise<any> {
-    return postData(
-        instance,
-        "/dataValueSets",
-        data,
-        _.pick(additionalParams, ["strategy", "dryRun"])
-    );
-}
-
-export async function postEventsData(
-    instance: Instance,
-    data: object,
-    additionalParams?: DataImportParams
-): Promise<any> {
-    return postData(instance, "/events", data, _.pick(additionalParams, ["dryRun"]));
-}
-
-export function buildMetadataDictionary(metadataPackage: Record<string, any[]>) {
-    return _(metadataPackage)
-        .values()
-        .flatten()
-        .tap(array => {
-            const dataSetElements = _.flatten(
-                _.map(metadataPackage.dataSets ?? [], e =>
-                    _.map(e.dataSetElements ?? [], ({ dataElement }) => dataElement)
-                )
-            );
-
-            const groupDataElements = _.flatten(
-                _.map(metadataPackage.dataElementGroups ?? [], e => e.dataElements ?? [])
-            );
-
-            const groupSetDataElements = _.flatten(
-                _.map(metadataPackage.dataElementGroupSets ?? [], e =>
-                    _.flatten(_.map(e.dataElementGroups ?? [], ({ dataElements }) => dataElements))
-                )
-            );
-
-            array.push(...dataSetElements, ...groupDataElements, ...groupSetDataElements);
-        })
-        .keyBy("id")
-        .value();
-}
 
 export async function requestJSONDownload(
     SyncClass: SyncronizationClass,
